+++ CHANGELOG.md
@@ -6,6 +6,7 @@ It is based on Tidepool Blip 1.27.
 ### Fixed
 - YLP-842 Patient user click logo leads to a blank page
 - YLP-851 Translation error in Loop Mode widget (DE)
+- YLP-852 App crash in the daily view on mobile - on resize display
 ### Engineering Use
 - YLP-813 Add TU for cookies manager